services/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0
python-multipart==0.0.6  # для форм